+++ src/main/java/post/PostUrlService.java
@@ -1,13 +1,16 @@
 package post;
 
+import account.AccountType;
 import board.Board;
 import board.BoardRepository;
 import customException.InvalidValueException;
 import customException.NoExistBoardException;
 import customException.NoExistParameterException;
 import customException.NoExistPostException;
+import customException.NotAllowedAuthorityException;
 import url.Request;
 import url.Response;
+import url.Session;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -31,8 +34,11 @@ public class PostUrlService {
         return post;
     }
 
-    public Response add(Request request) throws IOException, NoExistParameterException, InvalidValueException, NoExistBoardException {
+    public Response add(Request request) throws IOException, NoExistParameterException, InvalidValueException, NoExistBoardException, NotAllowedAuthorityException {
         try {
+            if (!request.getSession().isSigned()) {
+                throw new NotAllowedAuthorityException();
+            }
             System.out.println("게시글을 작성합니다.");
             Long boardId = Long.parseLong(request.getParameter("boardId"));
             Board board = boardRepository.get(boardId);
@@ -68,10 +74,15 @@ public class PostUrlService {
         }
     }
 
-    public Response remove(Request request) throws InvalidValueException, NoExistParameterException {
+    public Response remove(Request request) throws InvalidValueException, NoExistParameterException, NotAllowedAuthorityException {
         try {
             Long postId = Long.parseLong(request.getParameter("postId"));
-            Long boardId = findPost(postId).getBoardId();
+            Post post = findPost(postId);
+            if (hasAuthority(request)) {
+                throw new NotAllowedAuthorityException();
+            }
+
+            Long boardId = post.getBoardId();
             postRepository.remove(postId);
             Board board = boardRepository.get(boardId);
             board.removePost(postId);
@@ -83,10 +94,13 @@ public class PostUrlService {
         }
     }
 
-    public Response edit(Request request) throws IOException, NoExistPostException, InvalidValueException, NoExistParameterException {
+    public Response edit(Request request) throws IOException, NoExistPostException, InvalidValueException, NoExistParameterException, NotAllowedAuthorityException {
         try {
             Long postId = Long.parseLong(request.getParameter("postId"));
             Post post = findPost(postId);
+            if (hasAuthority(request)) {
+                throw new NotAllowedAuthorityException();
+            }
             System.out.print("제목을 입력해 주십시오. : ");
             String title = br.readLine();
             System.out.print("내용을 입력해 주십시오. : ");
@@ -98,5 +112,15 @@ public class PostUrlService {
         } catch (NumberFormatException e) {
             throw new InvalidValueException("postId", e);
         }
+    }
+
+    private boolean hasAuthority(Request request) throws NoExistParameterException, NoExistPostException {
+        Session session = request.getSession();
+        Post post = findPost(Long.parseLong(request.getParameter("postId")));
+        if (!session.isSigned()) return false;
+        if (!session.getAccountType().equals(AccountType.ADMIN) && !session.getSessionId().equals(post.getWriterId()))
+            return false;
+
+        return true;
     }
 }
